Cleans syntax via Rubocop
Much of the code in the repo did not follow syntax conventions.
This commit cleans out warnings given by Rubocop down to a
minimal number. If syntax overrides are required, explicit
exceptions can be made later on.

A few of the major issues are not covered here. Currently, there
are 8 offenses left untreated. These would require deeper-level
decisions to be made about the behaviour of the application,
and should be addressed separately.